+++ src/main/java/devmountain/group2/entities/RatingEntity.java
@@ -1,4 +1,4 @@
-package devmountain.group2;
+package devmountain.group2.entities;
 
 import jakarta.persistence.*;
 import lombok.Getter;
