+++ src/main/java/devmountain/group2/repositories/RatingRepository.java
@@ -1,5 +1,6 @@
-package devmountain.group2;
+package devmountain.group2.repositories;
 
+import devmountain.group2.entities.RatingEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
