+++ src/main/java/devmountain/group2/services/RatingService.java
@@ -1,4 +1,6 @@
-package devmountain.group2;
+package devmountain.group2.services;
+
+import devmountain.group2.dtos.CreateRatingPojo;
 
 public interface RatingService {
     void createDishRating(CreateRatingPojo createRatingPojo);
