+++ src/main/java/devmountain/group2/services/ReviewService.java
@@ -1,4 +1,6 @@
-package devmountain.group2;
+package devmountain.group2.services;
+
+import devmountain.group2.dtos.CreateReviewPojo;
 
 public interface ReviewService {
     void createReview(CreateReviewPojo createReviewPojo);
